alerts

+++ src/components/dashboard/sidebar.js
@@ -1,6 +1,5 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { FaUserCircle } from "react-icons/fa";
-import getTasks from "../../apicall/gettasks";
 import { RiLogoutBoxRFill } from "react-icons/ri";
 import { TiUserDelete } from "react-icons/ti";
 import Logout from "../../apicall/logout"
@@ -9,7 +8,6 @@ import { Redirect } from "react-router-dom";
 import Login from "../../components/login"
 import swal from "sweetalert";
 const Sidebar = (props) => {
-    const [tasks, setTasks] = useState();
     const [logout, setLogout] = useState({});
     const [del, setDel] = useState({})
     const handleClick = () => {
@@ -25,18 +23,23 @@ const Sidebar = (props) => {
 
     }
     const handleDelete = async () => {
-        swal()
-        setDel(await DeleteUser());
-        document.getElementById("root").style.marginRight = "0px";
+        const remove = await swal({
+            title: "Delete Account",
+            text: "Deleting your account will remove all information from our database this can not be undone!",
+            buttons: true,
+            dangerMode: true,
+        })
+
+        if (remove) {
+            swal("Your account has been removed!", {
+                icon: "success",
+            });
+            setDel(await DeleteUser())
+            document.getElementById("root").style.marginRight = "0px";
+        }
 
     }
-    useEffect(() => {
-        async function fetchtasks() {
-            let a = await getTasks();
-            setTasks(a)
-        };
-        fetchtasks()
-    }, [])
+
     return (
         <div className={props.open ? "Sidebar__open" : "Sidebar__close"}>
             {props.open ? <div className="Sidebar__closebutton" onClick={handleClick}>
@@ -59,6 +62,7 @@ const Sidebar = (props) => {
             </div>
             {/*eslint-disable-next-line*/}
             {logout.response == true ? <Redirect to="/" component={Login} /> : undefined}
+            {/*eslint-disable-next-line*/}
             {del.response == true ? <Redirect to="/" component={Login} /> : undefined}
 
         </div>
